feat(types): extract data model into oxc_coverage_types crate (#45)
PR B of issue #45's umbrella plan. Extracts the Istanbul-compatible
coverage data model into its own publishable crate so non-instrument
consumers (future reporters, source-map remappers, V8 converters in
this suite + downstream Rust JS tools) can depend on the types without
pulling in the instrumenter.

New crate: oxc_coverage_types v0.1.0
  - UnhandledPragma, FileCoverage, Location, Position, FnEntry,
    BranchEntry, parse_coverage_map
  - Private serde null-tolerant deserializer helpers
  - Pure data model, zero internal deps (just serde + serde_json)

oxc_coverage_instrument changes:
  - Depend on oxc_coverage_types via path + version specifier so
    cargo publish works through the standard release flow.
  - Re-export the same public types at the crate root via
    `pub use oxc_coverage_types::{...}`, preserving the existing
    public API surface byte-for-byte.
  - CoverageMaps + the from_maps builder (private, crate-internal)
    move to a new coverage_builder.rs module on the instrument side
    because their input shape (Vecs keyed by counter id) is a
    transform-pass implementation detail, not part of the data model.
    `FileCoverage::from_maps` (impl method) becomes
    `coverage_builder::build_file_coverage` (free function).
  - Internal callsites in instrument.rs, pragma.rs, source_maps.rs,
    transform.rs, v8_to_istanbul.rs switch from `use crate::types::...`
    to `use oxc_coverage_types::...`.

Release workflow (release-npm.yml):
  - Publish oxc_coverage_types first, then oxc_coverage_instrument
    (topological order so the lib's crates.io publish can resolve
    oxc_coverage_types from the registry).

Working name oxc_coverage_types: keeps the oxc prefix (safe under
fallow-rs ownership) with the _types suffix making the limited scope
honest. Final naming for the data-model crate is still open per #45's
naming refresh: candidate neutral names (istanbul_coverage,
istanbul_data) remain pending coordination with OJ Kwon on
istanbul-oxide interop. If that conversation ever happens and a
neutral name is preferred, oxc_coverage_types can be deprecated and
re-exported from the renamed crate; doing it now without the
conversation would preemptively claim the istanbul-* namespace.

Verified locally:
  - cargo check / test --workspace --all-targets (passes)
  - cargo clippy --workspace --all-targets -- -D warnings (passes)
  - cargo fmt --all --check (passes)
  - cargo doc --workspace --no-deps --document-private-items -D warnings (clean)
  - cargo deny check (passes)
  - typos (clean)
  - cargo publish -p oxc_coverage_types --dry-run (5 files, 15.3 KiB, packages clean)
  - napi build --release + node test.mjs (16/16 pass)
  - scripts/istanbul-upstream-specs.mjs (8/8 pass)
  - scripts/istanbul-diff.mjs (27/27 fixtures match)
  - scripts/real-world-parity.mjs (5/5 libs match)

Note for next release: this is the first release that needs to
publish oxc_coverage_types@0.1.0 alongside the lib bump. The workflow
publishes types first; if the crates.io index propagation is slow,
the lib publish may need a one-time manual rerun.

Refs #45
