+++ crates/oxc-coverage-instrument/Cargo.toml
@@ -15,6 +15,9 @@ categories = ["development-tools::testing", "web-programming"]
 workspace = true
 
 [dependencies]
+# Suite data-model crate. Path + version: path for workspace builds, version
+# for `cargo publish` so the lib crate can publish standalone from crates.io.
+oxc_coverage_types = { path = "../oxc-coverage-types", version = "0.1.0" }
 oxc_allocator = "0.126"
 oxc_ast = "0.126"
 oxc_codegen = "0.126"
